refactor event liseners
